First Commit
Heli Drop is a simple 2D XNA game that I created for my graphics class.
You will quickly notice that art and graphics are not my thing. Although
I didn't thouroughly enjoy creating this game, I did learn quite a bit.
The object of the game is to get rid of all the runners by hitting them
with your spotlight. See if you can get them all... it is possible.